requirements.txt
PyYAML~=6.0
pandas~=1.4.2
numpy~=1.22.4
scikit-learn~=1.1.1
pytest~=7.1.2